Fixed min_p_threshold in enrichment_in_abundance
The min_p_threshold parameter was not working for the enrichment-comparison method in leapR. This is a simple fix that cuts down the output 'results' dataframe.
Please also check the documentation for the leapR function. The min_p_threshold parameter is described as being only for the enrichment-comparison method, but also works for the enrichment_in_pathway method.