Update enrichment_in_groups.R
Changes proposed to fix the following error from running enrichment_in_groups.R

Below is a brief description of the error, what is causing it and a proposed solution/fix.

Error
Error in if (class(enr) == "try-error") { : the condition has length > 1

Debugging (Problem identified)
See lines 77 and 78 in enrichment_in_groups.R (master branch as of March 1, 2023)
The error occurs beause length(class(enr)) is not 1, so the if statement doesn't get executed.
class(enr)
[1] "ks.test" "htest"
Besides, to follow good practice, there should not be a test on class(enr) in order to catch errors.

Proposed fix:
Replaced the if statement block of code with a tryCatch statement to handle potential errors and warnings more elegantly.